fix(skeletoncss) : skelton css common.css로 따로 빼서 사용
